refactor: AppFileStructure
